Remove min-h-screen from `ExerciseModal` (#604)
Was checking #539 and I noticed the ExerciseModal has changed from
sidebar to dialog. While it was a sidebar, it made sense to have
`min-h-screen` but with the current style, the `min-h-screen` creates
extra vertical space.

Removing the `min-h-screen` should fix it. 
- The dialog still have vertical scroll when the window's height is less
than the modal
- The UI still look proper in mobile screen too.


https://github.com/user-attachments/assets/9be06250-a62c-46ab-8f8f-6466cd555a4e



<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

* **Style**
* Updated modal styling to remove enforced minimum screen height,
allowing for more flexible modal sizing.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->